requirements.txt
asgiref==3.8.1
certifi==2023.7.22
charset-normalizer==3.1.0
Django==5.1-rc1
djangorestframework==3.14.0
idna==3.7
money==1.3.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
stripe==9.7.0
typing_extensions==4.6.2
urllib3==2.0.7